feature save color
