+++ Feature/Util.js
@@ -1,14 +1,7 @@
 export const $ = document.querySelector.bind(document);
 export const $$ = document.querySelectorAll.bind(document);
 
-export const buttonFeature = $(".button-feature-group");
 export const buttonActions = $(".button-actions");
-
-export const BUTTON_FEATURE = {
-  ROW: buttonFeature.children[0],
-  CELL: buttonFeature.children[1],
-  ERASER: buttonFeature.children[2],
-};
 
 export function AlertAfterClose(event) {
   event.preventDefault();
@@ -44,34 +37,6 @@ export function unActiveButtonSave(rowNode) {
   saveNode.src = "./Assets/Icons/create-unactive-icon.svg";
 }
 
-export function activeButtonCreate(rowNode) {
-  if (!rowNode) return;
-  const createNode = rowNode.querySelector(".save-new-record");
-  createNode.classList.add("active");
-  createNode.src = "./Assets/Icons/create-active-icon.svg";
-}
-export function unActiveButtonCreate(rowNode) {
-  if (!rowNode) return;
-  const createNode = rowNode.querySelector(".save-new-record.active");
-  if (!createNode) return;
-  createNode.classList.remove("active");
-  createNode.src = "./Assets/Icons/create-unactive-icon.svg";
-}
-
-export function activeButtonEraserNewNode(rowNode) {
-  if (!rowNode) return;
-  const eraserNewNode = rowNode.querySelector(".clear-new-node");
-  eraserNewNode.classList.add("active");
-  eraserNewNode.src = "./Assets/Icons/eraser-active-icon.svg";
-}
-
-export function unActiveButtonEraserNewNode(rowNode) {
-  if (!rowNode) return;
-  const eraserNewNode = rowNode.querySelector(".clear-new-node.active");
-  if (!eraserNewNode) return;
-  eraserNewNode.classList.remove("active");
-  eraserNewNode.src = "./Assets/Icons/eraser-unactive-icon.svg";
-}
 export function activeButtonEraserInput(node) {
   if (!node) return;
   const eraserNode = node.querySelector("img.eraser");
